+++ frontend/appflowy_flutter/packages/flowy_infra_ui/lib/src/flowy_overlay/flowy_dialog.dart
@@ -6,19 +6,22 @@ const overlayContainerMaxWidth = 760.0;
 const overlayContainerMinWidth = 320.0;
 
 class FlowyDialog extends StatelessWidget {
-  final Widget? title;
-  final ShapeBorder? shape;
-  final Widget child;
-  final BoxConstraints? constraints;
-  final EdgeInsets padding;
   const FlowyDialog({
+    super.key,
     required this.child,
     this.title,
     this.shape,
     this.constraints,
     this.padding = _overlayContainerPadding,
-    Key? key,
-  }) : super(key: key);
+    this.backgroundColor,
+  });
+
+  final Widget? title;
+  final ShapeBorder? shape;
+  final Widget child;
+  final BoxConstraints? constraints;
+  final EdgeInsets padding;
+  final Color? backgroundColor;
 
   @override
   Widget build(BuildContext context) {
@@ -26,7 +29,7 @@ class FlowyDialog extends StatelessWidget {
     final size = windowSize * 0.7;
     return SimpleDialog(
         contentPadding: EdgeInsets.zero,
-        backgroundColor: Theme.of(context).cardColor,
+        backgroundColor: backgroundColor ?? Theme.of(context).cardColor,
         title: title,
         shape: shape ??
             RoundedRectangleBorder(borderRadius: BorderRadius.circular(8)),
